darwin_stop_world.c: Update for -m64 multilib.
        * darwin_stop_world.c: Update for -m64 multilib.
        * include/private/gcconfig.h: Likewise.
        * powerpc_darwin_mach_dep.s: Likewise.

From-SVN: r96968

+++ boehm-gc/darwin_stop_world.c
@@ -1,5 +1,7 @@
 #include "private/pthread_support.h"
 
+/* This probably needs more porting work to ppc64. */
+
 # if defined(GC_DARWIN_THREADS)
 
 /* From "Inside Mac OS X - Mach-O Runtime Architecture" published by Apple
@@ -14,13 +16,12 @@
 */
 #define PPC_RED_ZONE_SIZE 224
 
-/* Not 64-bit clean. Wait until Apple defines their 64-bit ABI */
 typedef struct StackFrame {
-  unsigned int	savedSP;
-  unsigned int	savedCR;
-  unsigned int	savedLR;
-  unsigned int	reserved[2];
-  unsigned int	savedRTOC;
+  unsigned long	savedSP;
+  unsigned long	savedCR;
+  unsigned long	savedLR;
+  unsigned long	reserved[2];
+  unsigned long	savedRTOC;
 } StackFrame;
